requirements.txt
Django==4.2.6
python-telegram-bot==13.2
python-dotenv==1.0.0